# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basketball Performance Tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface Player {
  id: number;
  name: string;
  team: string;
  position: string;
  number: number;
  season: string;
  createdAt: string;
}

export interface CreatePlayerBody {
  name: string;
  team: string;
  position: string;
  number: number;
  season: string;
}

export interface Game {
  id: number;
  date: string;
  opponent: string;
  homeAway: string;
  result: string;
  teamScore: number;
  opponentScore: number;
  createdAt: string;
}

export interface CreateGameBody {
  date: string;
  opponent: string;
  homeAway: string;
  result: string;
  teamScore: number;
  opponentScore: number;
}

export interface GameStat {
  id: number;
  playerId: number;
  gameId: number;
  points: number;
  assists: number;
  rebounds: number;
  steals: number;
  blocks: number;
  turnovers: number;
  minutesPlayed: number;
  fieldGoalsMade: number;
  fieldGoalsAttempted: number;
  threesMade: number;
  threesAttempted: number;
  freeThrowsMade: number;
  freeThrowsAttempted: number;
  plusMinus: number;
  game?: Game;
}

export interface CreateStatBody {
  playerId: number;
  gameId: number;
  points: number;
  assists: number;
  rebounds: number;
  steals: number;
  blocks: number;
  turnovers: number;
  minutesPlayed: number;
  fieldGoalsMade: number;
  fieldGoalsAttempted: number;
  threesMade: number;
  threesAttempted: number;
  freeThrowsMade: number;
  freeThrowsAttempted: number;
  plusMinus: number;
}

export interface PlayerAnalytics {
  playerId: number;
  gamesPlayed: number;
  ppg: number;
  apg: number;
  rpg: number;
  spg: number;
  bpg: number;
  tpg: number;
  mpg: number;
  fieldGoalPct: number;
  threePointPct: number;
  freeThrowPct: number;
  efficiencyRating: number;
  plusMinusAvg: number;
  consistencyScore: number;
  fatigueIndex: number;
  last5Ppg: number;
  last5Apg: number;
  last5Rpg: number;
  seasonTotalPoints: number;
  seasonTotalAssists: number;
  seasonTotalRebounds: number;
}

export interface TrendPoint {
  gameId: number;
  date: string;
  opponent: string;
  points: number;
  assists: number;
  rebounds: number;
  efficiencyRating: number;
  plusMinus: number;
}

export interface LeaderboardEntry {
  playerId: number;
  playerName: string;
  team: string;
  position: string;
  ppg: number;
  apg: number;
  rpg: number;
  efficiencyRating: number;
  gamesPlayed: number;
}

export interface TeamSummary {
  totalPlayers: number;
  totalGames: number;
  avgPpg: number;
  avgApg: number;
  avgRpg: number;
  topScorer: string;
  topAssister: string;
  topRebounder: string;
}