# lib/api-zod/src/generated/types/createGameBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basketball Performance Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGameBody {
  date: string;
  opponent: string;
  homeAway: string;
  result: string;
  teamScore: number;
  opponentScore: number;
}