# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Basketball Performance Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface Game {
  id: number;
  date: string;
  opponent: string;
  homeAway: string;
  result: string;
  teamScore: number;
  opponentScore: number;
  createdAt: string;
}